Fix a bug in smi_core.SingleMoleculeData class usage of ConnectID field

+++ MATLAB/+smi_core/@SingleMoleculeData/SingleMoleculeData.m
@@ -87,7 +87,7 @@ classdef SingleMoleculeData
             SMD.FrameNum=[];
             SMD.PValue=[];
             SMD.LogLikelihood=[];
-            SMD.ConnectID=[];
+            SMD.ConnectID=uint32([]);
             SMD.IndSMD={};
             SMD.ThreshFlag=[];
             SMD.DriftX=[];
